refact: language processing chatbot
