+++ bot-lib/languageprocessing/chatbot.py
@@ -1,34 +1,11 @@
-from transformers import AutoTokenizer  
-from transformers import AutoModel 
-import torch
 import numpy as np
-import pandas as pd
 
-class QuestionEmbeddings():
-    def __init__(self, question_path, no_answer, similarity_threshold=0.8):
-        self.no_answer = no_answer
+class Chatbot():
+    """Class that generate ADA's answer considering set of frequent questions. 
+    """
+    def __init__(self, question_embeddings, similarity_threshold=0.8):
         self.similarity_threshold = similarity_threshold
-        self.tokenizer = AutoTokenizer.from_pretrained('./model', do_lower_case=False)
-        self.model = AutoModel.from_pretrained('./model')
-        perguntas_frequentes_int = pd.read_csv(question_path, sep = ';')
-        self.perguntas_frequentes = self.get_database_embs(perguntas_frequentes_int)
-
-    def get_database_embs(self, perguntas_frequentes):
-        perguntas_frequentes['Sentence Embedding'] = perguntas_frequentes['PERGUNTAS'].apply(self.get_sentence_embs)
-        perguntas_frequentes['SAUDACAO'] = perguntas_frequentes['SAUDACAO'].astype(int)
-        return perguntas_frequentes
-
-    def get_embs_bertinbau(self, frase):
-        input_ids = self.tokenizer.encode(frase, return_tensors='pt')
-        with torch.no_grad():
-            outs = self.model(input_ids)
-            encoded = outs[0][0, 1:-1] 
-        return encoded
-
-    def get_sentence_embs(self, frase):
-        frase_embs = self.get_embs_bertinbau(frase.lower()).numpy()
-        final = np.mean(frase_embs, axis=0)
-        return final
+        self.question_embeddings = question_embeddings
     
     def semelhanca_cossenos(self, a,b):
         return np.dot(a,b)/(np.linalg.norm(a)*np.linalg.norm(b))
@@ -37,8 +14,9 @@ class QuestionEmbeddings():
         mais_similar = ""
         maior_score_similaridade = -999
 
-        sentence_emb = self.get_sentence_embs(frase)
-        for i, row in self.perguntas_frequentes.iterrows():
+        sentence_emb = self.question_embeddings.get_sentence_embs(frase)
+        perguntas_frequentes_emb = self.question_embeddings.get_perguntas_frequentes()
+        for i, row in perguntas_frequentes_emb.iterrows():
             similaridade = self.semelhanca_cossenos(sentence_emb,row['Sentence Embedding'])
             if similaridade > maior_score_similaridade:
                 maior_score_similaridade = similaridade
